Add files via upload
Naive Bayes classification is a probabilistic machine learning algorithm used for classification tasks, particularly text classification. 
It's based on Bayes' Theorem, which calculates the probability of a hypothesis given the evidence. 
The "naive" part refers to the algorithm's assumption that features are independent of each other, meaning the presence of one feature doesn't affect the presence of another.