Add files via upload
Hierarchical clustering is a data analysis technique that organizes data points into a hierarchy of clusters, visualized as a tree-like structure called a dendrogram. 
It builds this hierarchy either by starting with individual data points and merging them into larger clusters (bottom-up or agglomerative approach) or by starting with a single cluster of all data and splitting it into smaller clusters (top-down or divisive approach).